refactor: match mock date in tests as per topics, split tests
