+++ packages/author-profile/__tests__/android/author-profile-head.test.js
@@ -1,0 +1,5 @@
+import shared from "../author-profile-head";
+
+describe("AuthorProfileHead tests on android", () => {
+  shared();
+});
