+++ packages/author-profile/__tests__/author-profile-head.js
@@ -1,0 +1,63 @@
+import React from "react";
+import { shallow } from "enzyme";
+import { AuthorProfileHead as AuthorProfileHeadWithoutTracking } from "../src/author-profile-head";
+import AuthorProfileHeadTwitter from "../src/author-profile-head-twitter";
+
+export default () => {
+  const headProps = {
+    twitter: "testTwitter",
+    isLoading: false
+  };
+
+  it("should not re-render when twitter is changed", () => {
+    const wrapper = shallow(
+      <AuthorProfileHeadWithoutTracking {...headProps} />
+    );
+
+    expect(wrapper.find("AuthorProfileHeadTwitter")).toMatchSnapshot(
+      "1. Should not re-render when twitter prop is changed"
+    );
+
+    wrapper.setProps({
+      twitter: "newTestTwitter"
+    });
+
+    expect(wrapper.find("AuthorProfileHeadTwitter")).toMatchSnapshot(
+      "2. Should not re-render when twitter prop is changed"
+    );
+  });
+
+  it("should only re-render when isLoading is changed", () => {
+    const wrapper = shallow(
+      <AuthorProfileHeadWithoutTracking {...headProps} />
+    );
+
+    expect(wrapper).toMatchSnapshot(
+      "3. Should re-render when isLoading prop is changed"
+    );
+
+    wrapper.setProps({
+      isLoading: true
+    });
+
+    expect(wrapper).toMatchSnapshot(
+      "4. Should re-render when isLoading prop is changed"
+    );
+  });
+
+  const mockOnPress = jest.fn();
+  const twitterProps = {
+    isLoading: false,
+    onTwitterLinkPress: mockOnPress,
+    twitter: "testTwitterHandle",
+    url: "https://twitter.com/testTwitterHandle"
+  };
+
+  it("should handle the twitter link when pressed", () => {
+    const wrapper = shallow(<AuthorProfileHeadTwitter {...twitterProps} />);
+
+    wrapper.find("TextLink").simulate("press");
+
+    expect(mockOnPress).toHaveBeenCalled();
+  });
+};
