+++ packages/author-profile/__tests__/author-profile.js
@@ -1,12 +1,9 @@
-/* global context */
 import React from "react";
 import renderer from "react-test-renderer";
-import { shallow } from "enzyme";
+import mockDate from "mockdate";
 import { fixtureGenerator } from "@times-components/provider-test-tools";
 import { delay, MockedProvider } from "@times-components/utils";
 import AuthorProfile from "../src/author-profile";
-import AuthorProfileHead from "../src/author-profile-head"; // eslint-disable-line import/no-named-as-default
-import AuthorProfileHeadTwitter from "../src/author-profile-head-twitter";
 import longSummaryLength from "../author-profile-constants";
 
 export default () => {
@@ -44,10 +41,12 @@ export default () => {
         resolvedOptions: () => ({ timeZone: "Europe/London" })
       })
     };
+    mockDate.set(1514764800000, 0);
   });
 
   afterAll(() => {
     global.Intl = realIntl;
+    mockDate.reset();
   });
 
   it("should render with images", async () => {
@@ -96,44 +95,6 @@ export default () => {
     );
   });
 
-  context("AuthorProfileHead shared tests", () => {
-    const mockOnPress = jest.fn();
-    const twitterProps = {
-      isLoading: false,
-      onTwitterLinkPress: mockOnPress,
-      twitter: "testTwitterHandle",
-      url: "www.twitter.com/"
-    };
-
-    it("should render with no twitter handle", () => {
-      const wrapper = shallow(
-        <AuthorProfileHead {...twitterProps} twitter="sdasdasd" />
-      );
-
-      expect(wrapper).toMatchSnapshot(
-        "5. Render an author profile header with a twitter link"
-      );
-    });
-
-    it("should render with no twitter handle", () => {
-      const wrapper = shallow(
-        <AuthorProfileHead {...twitterProps} twitter="" />
-      );
-
-      expect(wrapper).toMatchSnapshot(
-        "6. Render an author profile header without a twitter link"
-      );
-    });
-
-    it("should handle the twitter link when pressed", () => {
-      const wrapper = shallow(<AuthorProfileHeadTwitter {...twitterProps} />);
-
-      wrapper.find("TextLink").simulate("press");
-
-      expect(mockOnPress).toHaveBeenCalled();
-    });
-  });
-
   it("should send analytics when rendering an author profile page", () => {
     const reporter = jest.fn();
 
@@ -144,13 +105,7 @@ export default () => {
     );
 
     const call = reporter.mock.calls[0][0];
-    const callWithoutEventTime =
-      call.attrs && call.attrs.eventTime
-        ? [{ ...call, attrs: { ...call.attrs, eventTime: null } }]
-        : call;
 
-    expect(callWithoutEventTime).toMatchSnapshot(
-      "8. Author profile page analytics"
-    );
+    expect(call).toMatchSnapshot("5. Author profile page analytics");
   });
 };
