+++ packages/author-profile/__tests__/ios/author-profile-head.test.js
@@ -1,0 +1,5 @@
+import shared from "../author-profile-head";
+
+describe("AuthorProfileHead tests on ios", () => {
+  shared();
+});
