+++ packages/author-profile/__tests__/ios/author-profile.test.js
@@ -1,4 +1,4 @@
-import shared from "../shared";
+import shared from "../author-profile";
 
 describe("AuthorProfile tests on ios", () => {
   shared();
