+++ packages/author-profile/__tests__/web/author-profile-head.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../author-profile-head";
+
+describe("AuthorProfileHead tests on web", () => {
+  shared();
+});
