+++ packages/author-profile/package.json
@@ -44,6 +44,7 @@
     "enzyme": "3.3.0",
     "eslint": "4.9.0",
     "jest": "21.2.1",
+    "mockdate": "2.0.2",
     "prettier": "1.8.2",
     "react": "16.3.1",
     "react-dom": "16.3.1",
